Adds Jasmine tests (run using npm test) - moves date related fuctions into displayHelper file - test displayHelper functions
