+++ source/javascripts/all.js
@@ -3,17 +3,8 @@ var L = require('leaflet');
 var $ = require('jquery');
 var tj = require('./togeojson');
 var numberIcon = require('./leaflet_awesome_number_markers');
+var displayHelper = require('./displayHelper');
 
-var getNowYMD = function(dt){
-    var y = dt.getFullYear();
-    var m = ("00" + (dt.getMonth()+1)).slice(-2);
-    var d = ("00" + dt.getDate()).slice(-2);
-    var hh = ("00" + dt.getHours()).slice(-2);
-    var mm =  ("00" + dt.getMinutes()).slice(-2);
-    var result = y + "年" + m + "月" + d + "日" + hh + "時" + mm + "分";
-    return result;
-  };
-  
 $(function(){
     // MIERUNEMAPのAPIキーはローカル環境では表示されないのでご注意(https://codeforjapan.github.io/mapprint/　でのみ表示される）
     // サーバ上の場合のみMIERUNE地図を使う
@@ -34,7 +25,7 @@ $(function(){
 
     $('#date').text(() => {
       const d = new Date()
-      return `このマップは ${d.getFullYear()}年${d.getMonth()+1}月${d.getDate()}日 ${d.getHours()}:${d.getMinutes()} に印刷しました。`;
+      return displayHelper.getPrintDate(d);
     });
 
     // 説明の表示/非表示
@@ -49,7 +40,7 @@ $(function(){
 
     $.ajax('./images/water-supply.kml').done(function (data, textStatus, jqXHR) {
         // データの最終更新日を表示（ローカルでは常に現在時刻となる）
-        var date = getNowYMD(new Date(jqXHR.getResponseHeader('date')));
+        var date = displayHelper.getNowYMD(new Date(jqXHR.getResponseHeader('date')));
         console.log(date);
         $('#datetime').html(date.toString());
         console.log(data);
